src/requirements.txt
av==13.1.0
hydra-core==1.3.2
lightning==2.4.0
lightning_utilities==0.11.9
moviepy==2.1.1
numpy==2.1.3
omegaconf==2.3.0
rich==13.9.4
rootutils==1.0.7
torch==2.5.1
torchmetrics==1.6.0
torchvision==0.20.1
tqdm==4.67.1
transformers==4.46.3
ultralytics==8.3.36
wandb==0.18.7
